Refactor transport: move things into raw and retransmit
