+++ examples/platform/nrf528xx/app/Service.cpp
@@ -41,7 +41,7 @@
 #include <support/ErrorStr.h>
 #include <system/SystemPacketBuffer.h>
 #include <transport/SecureSessionMgr.h>
-#include <transport/UDP.h>
+#include <transport/raw/UDP.h>
 
 #if CHIP_ENABLE_OPENTHREAD
 #include <openthread/message.h>
