+++ examples/wifi-echo/server/esp32/main/DataModelHandler.cpp
@@ -22,7 +22,7 @@
 
 #include "esp_log.h"
 #include <system/SystemPacketBuffer.h>
-#include <transport/MessageHeader.h>
+#include <transport/raw/MessageHeader.h>
 
 #include "DataModelHandler.h"
 #include "LEDWidget.h"
