+++ src/app/util/types_stub.h
@@ -1952,7 +1952,7 @@ typedef struct
  * ChipNodeId that's compatible.
  */
 #ifdef __cplusplus
-#include <transport/MessageHeader.h>
+#include <transport/raw/MessageHeader.h>
 static_assert(sizeof(chip::NodeId) == sizeof(uint64_t), "Unexpected node if size");
 // Make it easier to have unified function declarations across C and C++ source
 // files.
