+++ src/transport/PeerConnectionState.h
@@ -22,8 +22,8 @@
 #ifndef PEER_CONNCTION_STATE_H_
 #define PEER_CONNCTION_STATE_H_
 
-#include <transport/MessageHeader.h>
-#include <transport/PeerAddress.h>
+#include <transport/raw/MessageHeader.h>
+#include <transport/raw/PeerAddress.h>
 #include <transport/SecureSession.h>
 
 namespace chip {
