+++ src/transport/raw/BLE.cpp
@@ -24,8 +24,8 @@
 
 #include <support/CodeUtils.h>
 #include <support/logging/CHIPLogging.h>
-#include <transport/BLE.h>
-#include <transport/MessageHeader.h>
+#include <transport/raw/BLE.h>
+#include <transport/raw/MessageHeader.h>
 
 #include <inttypes.h>
