+++ src/transport/raw/BUILD.gn
@@ -15,23 +15,27 @@
 import("//build_overrides/chip.gni")
 import("${chip_root}/src/ble/ble.gni")
 
-static_library("transport") {
-  output_name = "libTransportLayer"
+static_library("raw") {
+  output_name = "libRawTransport"
 
   sources = [
-    "PeerConnectionState.h",
-    "PeerConnections.h",
-    "RendezvousParameters.h",
-    "RendezvousSession.cpp",
-    "RendezvousSession.h",
-    "RendezvousSessionDelegate.h",
-    "SecurePairingSession.cpp",
-    "SecurePairingSession.h",
-    "SecureSession.cpp",
-    "SecureSession.h",
-    "SecureSessionMgr.cpp",
-    "SecureSessionMgr.h",
+    "Base.h",
+    "MessageHeader.cpp",
+    "MessageHeader.h",
+    "PeerAddress.h",
+    "TCP.cpp",
+    "TCP.h",
+    "Tuple.h",
+    "UDP.cpp",
+    "UDP.h",
   ]
+
+  if (chip_config_network_layer_ble) {
+    sources += [
+      "BLE.cpp",
+      "BLE.h",
+    ]
+  }
 
   public_deps = [
     "${chip_root}/src/crypto",
@@ -39,6 +43,5 @@ static_library("transport") {
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
     "${chip_root}/src/platform",
-    "${chip_root}/src/transport/raw",
   ]
 }
