+++ src/transport/retransmit/tests/BUILD.gn
@@ -29,7 +29,7 @@ chip_test_suite("tests") {
   ]
 
   public_deps = [
-    "${chip_root}/src/retransmit",
+    "${chip_root}/src/transport/retransmit",
     "${nlio_root}:nlio",
     "${nlunit_test_root}:nlunit-test",
   ]
