+++ src/transport/tests/TestMessageHeader.cpp
@@ -27,7 +27,7 @@
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
 #include <support/TestUtils.h>
-#include <transport/MessageHeader.h>
+#include <transport/raw/MessageHeader.h>
 
 #include <nlunit-test.h>
